Add filmography to PersonDelegate
A person's filmography can be explored now.
The XmlListModels cannot be reordered so,
since the API returns the filmography ordered
by movie, it's not possible to group the results
by role without having repeated roles. The
current workaround shows the results grouped
by movie, with as many entries as roles the
person did on it.