Upload movie showtimes fetching
Now, the theaterListModel is exposed into the QML
context from the beginning (even if it's empty),
and the list view is connected / disconnected from the
model when fetching or if no results are found.

Besides, now the model is correctly cleaned up,
and no extra connections are created for the
webview.