Update MovieView to allow "cold-loading"
Favorite movies are currently saving just the movie identifier. In order
allow loading them using the MovieView, it has been updated so that it no
longer expects the "lightweight" object coming from the SearchView or the
BrowseView. Now, it's possible to only load a movie from its identifier.

When that's the case, no content is displayed until the movie is fully
fetched.