Workaround for the implicit index property in ListDelegate
Since I'm emitting the clicked signal with an index parameter to be able to
swap between delegates in the SearchView, and that index parameter is an
implicit one (so it's not available when used without a model), this
workaround prevents it from failing. It should be fixed so the signal
doesn't carry the index, though.